# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: QuizzGame API — Practice, Thi thử, Admin & Ngân hàng câu hỏi
  version: "1.3.0"
  description: >-
    Contract cho module Ôn tập (Practice), module Thi thử (Exam/Mock Exam),
    các endpoint Admin liên quan (quản lý câu hỏi, xử lý báo cáo câu hỏi,
    quản lý đề thi thử) và Ngân hàng câu hỏi (Question Bank).
    Khớp với `docs/FEATURE_LOG.md` Section 4 (Practice Module), Section 5
    (Admin Dashboard – Quản lý báo cáo câu hỏi), Section 6 (Exam Module –
    Thi thử) và Section 7 (Ngân hàng câu hỏi).

    Các endpoint khác của hệ thống (Auth, Onboarding, Points, Leaderboard...)
    chưa được mô tả trong file này.
servers:
  - url: http://localhost:4000
    description: Local dev
...
security: []
paths:
...
  /api/admin/questions:
    get:
      tags: [Admin]
      summary: Danh sách câu hỏi (phân trang, lọc)
      security: [{ AdminSecret: [] }]
      parameters:
        - name: subject
          in: query
          schema: { $ref: '#/components/schemas/SubjectId' }
        - name: difficulty
          in: query
          schema: { type: integer, enum: [1, 2, 3] }
        - name: isActive
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: Danh sách câu hỏi
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/QuestionFullDto' }
                  total: { type: integer, example: 150 }
                  page: { type: integer, example: 1 }
                  limit: { type: integer, example: 20 }
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
    post:
      tags: [Admin]
      summary: Tạo 1 câu hỏi
      security: [{ AdminSecret: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionInput'
      responses:
        '201':
          description: Câu hỏi đã tạo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionFullDto'
        '400':
          description: Body không hợp lệ (xem `details`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
...
  /api/admin/questions/{id}:
    put:
      tags: [Admin]
      summary: Cập nhật câu hỏi (partial update)
      security: [{ AdminSecret: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuestionInput'
            example:
              explanation: 'Giai thich da duoc cap nhat ro hon...'
      responses:
        '200':
          description: Câu hỏi sau khi cập nhật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionFullDto'
        '400':
          description: Body không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '404':
          description: questionId không tồn tại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Admin]
      summary: Ẩn câu hỏi (soft delete — set isActive=false)
      security: [{ AdminSecret: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Đã ẩn câu hỏi
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: 'Da an cau hoi thanh cong.' }
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '404':
          description: questionId không tồn tại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ───────────────── Admin: báo cáo câu hỏi (Admin Dashboard) ─────────────────
...
  /api/admin/exam-papers/{id}/questions:
    post:
      tags: [Admin]
      summary: Thêm 1 câu hỏi vào đề thi thử
      description: >-
        `options`/`correctAnswer` phải khớp `questionType` (kiểm tra bởi
        `validateQuestionShape()`): `MCQ_4` → `options` 4 chuỗi,
        `correctAnswer` số 0-3; `TRUE_FALSE_4` → `options` 4 chuỗi,
        `correctAnswer` 4 boolean (theo đúng thứ tự a/b/c/d); `FILL_BLANK` →
        không có `options`, `correctAnswer` mảng ≥1 chuỗi (so khớp sau
        `normalizeAnswer()`).
      security: [{ AdminSecret: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExamQuestionInput'
            examples:
              MCQ_4:
                summary: Trắc nghiệm 4 lựa chọn
                value:
                  chapter: 'Hàm số'
                  difficulty: 2
                  questionType: MCQ_4
                  points: 0.25
                  questionText: 'Hàm số y = x^3 - 3x đồng biến trên khoảng nào?'
                  options: ['(-1;1)', '(-∞;-1)', '(1;+∞)', '(-∞;-1) và (1;+∞)']
                  correctAnswer: 3
                  explanation: "y' = 3x^2 - 3 = 3(x-1)(x+1); y' > 0 khi x < -1 hoặc x > 1"
                  examYear: 2024
                  examCode: 'Mã đề 101'
              TRUE_FALSE_4:
                summary: 'Đúng/Sai 4 phát biểu (correctAnswer = 4 boolean)'
                value:
                  chapter: 'Hình học'
                  difficulty: 2
                  questionType: TRUE_FALSE_4
                  points: 1
                  questionText: >-
                    Cho hình chóp S.ABCD có đáy là hình vuông cạnh a, SA
                    vuông góc với đáy, SA = a. Xét tính đúng/sai của các
                    phát biểu sau:
                  options:
                    - 'a) SA vuông góc với BC'
                    - 'b) Góc giữa SC và mặt đáy bằng 45°'
                    - 'c) Thể tích khối chóp S.ABCD bằng a^3/3'
                    - 'd) SC = a√3'
                  correctAnswer: [true, true, false, false]
                  explanation: >-
                    a, b đúng theo tính chất hình chóp đều có cạnh bên vuông
                    góc đáy; c, d sai do tính toán thể tích/độ dài cạnh.
              FILL_BLANK:
                summary: 'Điền đáp án (correctAnswer = mảng đáp án chấp nhận)'
                value:
                  chapter: 'Đại số'
                  difficulty: 1
                  questionType: FILL_BLANK
                  points: 0.5
                  questionText: 'Giải phương trình 2x + 4 = 10. Đáp án: x = ___'
                  correctAnswer: ['x = 3', 'x=3', '3']
      responses:
        '201':
          description: Câu hỏi đã được tạo (`ExamQuestionFullDto`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamQuestionFullDto'
        '400':
          description: 'Lỗi schema Zod, hoặc `options`/`correctAnswer` không khớp `questionType`'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidBody:
                  value: { error: INVALID_REQUEST_BODY, message: 'Du lieu khong hop le' }
                invalidShape:
                  value: { error: EXAM_QUESTION_INVALID, message: "Du lieu cau hoi khong hop le voi dang 'MCQ_4'." }
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '404':
          description: '`id` (đề thi) không tồn tại'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: EXAM_PAPER_NOT_FOUND
                message: "Khong tim thay de thi thu '...'."
...
  /api/admin/exam-papers/{id}/questions/{qid}:
    patch:
      tags: [Admin]
      summary: Cập nhật 1 câu hỏi (partial update)
      description: >-
        ⚠️ Endpoint đã implement đầy đủ ở backend nhưng hiện **CHƯA có nơi
        nào ở frontend gọi tới** (`frontend/src/lib/api.ts` không có
        `adminUpdateExamQuestion`, `AdminExamPaperDetailPage` chỉ có nút
        "Xoá"). Xem `docs/FEATURE_LOG.md` Section 6 — Ghi chú kỹ thuật #11.
        Nếu request có `questionType`/`options`/`correctAnswer`, server
        validate lại toàn bộ shape theo dạng câu hỏi MỚI (lấy giá trị cũ cho
        field không đổi).
      security: [{ AdminSecret: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
        - name: qid
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExamQuestionInput'
            example:
              explanation: 'Giải thích đã được cập nhật rõ hơn...'
      responses:
        '200':
          description: Câu hỏi sau khi cập nhật (`ExamQuestionFullDto`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamQuestionFullDto'
        '400':
          description: 'Lỗi schema Zod (partial), hoặc sau khi áp input mới `options`/`correctAnswer` không khớp `questionType`'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidBody:
                  value: { error: INVALID_REQUEST_BODY, message: 'Du lieu khong hop le' }
                invalidShape:
                  value: { error: EXAM_QUESTION_INVALID, message: "Du lieu cau hoi khong hop le voi dang 'MCQ_4'." }
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '404':
          description: '`qid` không tồn tại hoặc không thuộc đề `id`'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: EXAM_QUESTION_NOT_FOUND
                message: "Khong tim thay cau hoi '...'."
    delete:
      tags: [Admin]
      summary: Ẩn 1 câu hỏi (soft delete — set isActive=false)
      description: >-
        Giữ lại bản ghi để không phá vỡ lịch sử chấm điểm các phiên đã làm
        câu này.
      security: [{ AdminSecret: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
        - name: qid
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Đã ẩn câu hỏi
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: 'Da an cau hoi thanh cong.' }
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '404':
          description: '`qid` không tồn tại hoặc không thuộc đề `id`'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: EXAM_QUESTION_NOT_FOUND
                message: "Khong tim thay cau hoi '...'."
components:
  securitySchemes:
    SessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session token nội bộ nhận được sau khi đăng nhập Firebase (xem Section 3 - Auth)
    AdminSecret:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Giá trị biến môi trường `ADMIN_SECRET`
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    AdminUnauthorized:
      description: Thiếu hoặc sai `X-Admin-Secret`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: ADMIN_UNAUTHORIZED
            message: 'Khong co quyen truy cap.'
  schemas:
    SubjectId:
      type: string
      enum: [TOAN, VAN, ANH, LY, HOA, SINH, SU, DIA, GDCD]
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Mã lỗi (dùng để ánh xạ HTTP status, xem `app.ts`)
        message:
          type: string
        details:
          type: array
          description: Chỉ có khi `error = INVALID_REQUEST_BODY` (lỗi Zod)
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
    QuestionPublicDto:
      type: object
      description: Câu hỏi gửi cho user khi bắt đầu/resume phiên — KHÔNG có correctAnswer/explanation
      properties:
        id: { type: string, format: uuid }
        subject: { $ref: '#/components/schemas/SubjectId' }
        chapter: { type: string, nullable: true }
        difficulty: { type: integer, enum: [1, 2, 3] }
        question: { type: string }
        options:
          type: array
          items: { type: string }
          minItems: 4
          maxItems: 4
    QuestionFullDto:
      allOf:
        - $ref: '#/components/schemas/QuestionPublicDto'
        - type: object
          properties:
            correctAnswer: { type: integer, minimum: 0, maximum: 3 }
            explanation: { type: string, nullable: true }
            examYear: { type: integer, nullable: true }
            examCode: { type: string, nullable: true }
            isActive: { type: boolean }
            createdAt: { type: string, format: date-time }
    CreateQuestionInput:
      type: object
      required: [subject, difficulty, question, options, correctAnswer]
      properties:
        subject: { $ref: '#/components/schemas/SubjectId' }
        chapter: { type: string, maxLength: 200 }
        difficulty: { type: integer, enum: [1, 2, 3] }
        question: { type: string, minLength: 5, maxLength: 2000 }
        options:
          type: array
          items: { type: string, minLength: 1 }
          minItems: 4
          maxItems: 4
        correctAnswer: { type: integer, minimum: 0, maximum: 3 }
        explanation: { type: string, maxLength: 2000 }
        examYear: { type: integer, minimum: 2000, maximum: 2100 }
        examCode: { type: string, maxLength: 50 }
    UpdateQuestionInput:
      allOf:
        - $ref: '#/components/schemas/CreateQuestionInput'
      description: Tất cả field đều tùy chọn (partial update)
...
    ExamQuestionType:
      type: string
      enum: [MCQ_4, TRUE_FALSE_4, FILL_BLANK]
...
    ExamQuestionOptions:
      description: >-
        4 lựa chọn (MCQ_4) hoặc 4 phát biểu (TRUE_FALSE_4); null với
        FILL_BLANK.
      nullable: true
      type: array
      items: { type: string }
    ExamCorrectAnswer:
      description: >-
        Đáp án đúng — kiểu dữ liệu phụ thuộc `questionType`: MCQ_4 là số
        nguyên 0-3, TRUE_FALSE_4 là mảng 4 boolean (theo đúng thứ tự options
        a/b/c/d), FILL_BLANK là mảng ≥1 chuỗi (các đáp án được chấp nhận, so
        khớp sau `normalizeAnswer()`).
      oneOf:
        - type: integer
          minimum: 0
          maximum: 3
        - type: array
          items: { type: boolean }
          minItems: 4
          maxItems: 4
        - type: array
          items: { type: string }
          minItems: 1
...
    ExamQuestionPublicDto:
      type: object
      description: >-
        Câu hỏi gửi cho học sinh khi bắt đầu phiên — ẨN
        `correctAnswer`/`explanation`.
      properties:
        id: { type: string, format: uuid }
        chapter: { type: string, nullable: true }
        difficulty: { type: integer, enum: [1, 2, 3] }
        questionType: { $ref: '#/components/schemas/ExamQuestionType' }
        points: { type: number, example: 0.25 }
        questionText: { type: string }
        options: { $ref: '#/components/schemas/ExamQuestionOptions' }
    ExamQuestionFullDto:
      allOf:
        - $ref: '#/components/schemas/ExamQuestionPublicDto'
        - type: object
          properties:
            examPaperId: { type: string, format: uuid }
            correctAnswer: { $ref: '#/components/schemas/ExamCorrectAnswer' }
            explanation: { type: string, nullable: true }
            examYear: { type: integer, nullable: true }
            examCode: { type: string, nullable: true }
            isActive: { type: boolean }
            createdAt: { type: string, format: date-time }
...
    CreateExamQuestionInput:
      type: object
      required: [difficulty, questionType, points, questionText, correctAnswer]
      properties:
        chapter: { type: string, maxLength: 200 }
        difficulty: { type: integer, enum: [1, 2, 3] }
        questionType: { $ref: '#/components/schemas/ExamQuestionType' }
        points: { type: number, exclusiveMinimum: 0, example: 0.25 }
        questionText: { type: string, minLength: 1, maxLength: 4000 }
        options:
          type: array
          items: { type: string, minLength: 1 }
          minItems: 4
          maxItems: 4
          description: 'Bắt buộc với MCQ_4/TRUE_FALSE_4, không gửi với FILL_BLANK'
        correctAnswer: { $ref: '#/components/schemas/ExamCorrectAnswer' }
        explanation: { type: string, maxLength: 4000 }
        examYear: { type: integer, minimum: 2000, maximum: 2100 }
        examCode: { type: string, maxLength: 50 }
    UpdateExamQuestionInput:
      allOf:
        - $ref: '#/components/schemas/CreateExamQuestionInput'
      description: Tất cả field đều tùy chọn (partial update)
...